Substantial improvements to the fake PAM library
The fake PAM library for testing now also supports a script-driven
test infrastructure that uses configuration files to specify the PAM
options, calls, flags, return status, prompts, responses, and logging
output and supports callbacks to check the PAM status before it's
closed.  See tests/fakepam/README for more information.

The fake PAM library now supports PAM_AUTHTOK, PAM_OLDAUTHTOK, and
PAM_CONV items and more correctly implements the PAM environment.  It
also supports intercepting getpwnam calls and returning a synthetic
structure, and intercepting Kerberos krb5_kuserok calls and using the
configured home directory to find .k5login.  Log output is now
returned as an array of strings rather than concatenating all the
messages together.  The fake PAM library now depends on the TAP
library and uses its facilities for allocating memory so that memory
allocation failures are reported with bail.

+++ tests/pam-util/vector-t.c
@@ -62,10 +62,9 @@ main(void)
     is_int(4, copy->count, "...and has right count");
     is_int(4, copy->allocated, "...and has right allocated count");
     for (i = 0; i < 4; i++) {
-        is_string(cstring, copy->strings[i], "...and string %lu is right",
-                  (unsigned long) i);
+        is_string(cstring, copy->strings[i], "...and string %d is right", i);
         ok(copy->strings[i] != vector->strings[i],
-           "...and pointer %lu is different", (unsigned long) i);
+           "...and pointer %d is different", i);
     }
     vector_free(copy);
     vector_clear(vector);
